requirements/requirements.txt
Flask==1.1.2
flask-paginate==0.7.0
flask-restplus==0.13.0
Werkzeug==0.16.1
requests==2.24.0
pymongo==3.10.1
lxml>=1.1.29
pandas==1.1.0
python-decouple==3.3
gunicorn==20.0.4